well, I think I found the cause of every performance problem in teh app. Just... don't know how to go about fixing it quite yet
