+++ source/SIL.AppBuilder.Portal.Frontend/src/data/containers/with-current-user/fetcher.tsx
@@ -190,8 +190,8 @@ export function withFetcher() {
     }
 
     return compose(
+      withTranslations,
       withData({}),
-      withTranslations
     )(WrapperClass);
   };
 }
