+++ source/SIL.AppBuilder.Portal.Frontend/src/data/containers/with-current-user/index.tsx
@@ -23,6 +23,7 @@ export const Provider = withFetcher()((props: ICurrentUserProps & { children: an
 
 export function withCurrentUserContext(InnerComponent) {
   return (props) => {
+    console.log('what is causing this to update?', props);
     return (
       <CurrentUserContext.Consumer>
         {(context) => <InnerComponent {...props} {...context} />}
