+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/users/list/index.tsx
@@ -6,7 +6,7 @@ import {
   withCurrentOrganization,
   IProvidedProps as WithCurrentOrgProps,
 } from '@data/containers/with-current-organization';
-import { withData } from '@ui/components/user-table/data';
+import { withData } from './user-table/data';
 import { withTranslations } from '@lib/i18n';
 
 import Display from './display';
