+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/users/list/user-table/row/multi-group-select/active-groups-display.tsx
