+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/users/list/user-table/row/multi-role-select/role-select.tsx
@@ -15,6 +15,7 @@ import {
 import { ROLE } from '@data/models/role';
 import { withUserRoles, IUserRoleProps } from '@data/containers/resources/user';
 import { compareVia } from '@lib/collection';
+import { withActiveRoles } from './with-active-roles';
 
 interface IOwnProps {
   roles: RoleResource[];
@@ -67,6 +68,7 @@ class RoleSelect extends React.Component<IProps> {
 }
 
 export default compose(
+  withActiveRoles,
   withProps((props: IOwnProps) => {
     const { user, organization } = props;
