+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/users/list/user-table/user-table.scss
